fix(race): use long type in the race class

+++ src/main/kotlin/racingcar/Race.kt
@@ -1,7 +1,7 @@
 package racingcar
 
 class Race(
-    val rounds: Int,
+    val rounds: Long,
     val cars: List<Car>,
 ) {
     fun run() {
